sun_server/requirements.txt
Django==1.8
Markdown==2.6.2
djangorestframework==3.1.1
